[Frontend] Improve readability of conditional passes (#1194)
**Context:** We are using the term `pipeline` a lot. We are also
defining compilation pipelines at the module level and modifying them
depending on configuration options.

**Description of the Change:** Change the compilation pipeline into
stages and improve readability of passes that are conditional on compile
time options.

**Benefits:** Readability.

---------
